--- src/main/java/com/datagi/teamcalendar/domain/announcement/repository/AnnouncementRepository.java
@@ -1,10 +1,0 @@
-package com.datagi.teamcalendar.domain.announcement.repository;
-
-import com.datagi.teamcalendar.domain.announcement.Announcement;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AnnouncementRepository extends JpaRepository<Announcement, Long> {
-
-}
